+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - perks managing for Fallout characters
 - dice rolls for skills checks for Fallout characters
 - equipment management for Fallout characters
+- attack damage calculations for Fallout characters
 
 ## [0.4.13] - 2026-03-04
 ### Modified